Factor out getTimeEntriesHandler
